requirements.txt
asgiref==3.2.10
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.2
django-bootstrap-form==3.4
django-crispy-forms==1.10.0
django-filter==2.4.0
django-import-export==2.4.0
et-xmlfile==1.0.1
gunicorn==20.0.4
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.5
Pillow==8.0.0
psycopg2==2.8.6
python-decouple==3.3
pytz==2020.1
PyYAML==5.3.1
sqlparse==0.3.1
static3==0.7.0
tablib==3.0.0
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
